requirements.txt
# requirements.txt file to install app dependencies
pandas
numpy
streamlit
scikit-learn
transformers
torch
newspaper3k
requests
python-dotenv